refactor(pack): split encode.rs into focused sub-modules
Split the 2033-line encode.rs into a directory module with clear
separation of concerns:

- mod.rs: PackEncoder struct, constructors, encode() dispatch,
  inner_encode() orchestrator (~500 lines)
- header.rs: pack wire-format encoding (header, offset, zlib)
- sort.rs: magic_sort() ordering + similarity pre-filters
- delta_search.rs: DeltaWindowEntry + try_as_offset_delta()
  with cfg-gated Rabin/Myers dual-engine
- parallel.rs: parallel_encode() Rayon path
- output.rs: encode_and_output_to_files() file I/O
- tests.rs: all 15 tests extracted from inline

No behavioral changes. All 15 tests pass, clippy clean.
