+++ src/internal/pack/encode/mod.rs
@@ -1,0 +1,497 @@
+//! Streaming Git pack encoder.
+//!
+//! The encoder has two modes:
+//!
+//! - `window_size == 0`: encode objects independently with zlib. Object compression is
+//!   parallelized with Rayon and input order is preserved.
+//! - `window_size > 0`: group and sort objects, search a sliding window for a suitable delta
+//!   base, encode the target as an offset delta when profitable, then zlib-compress the payload.
+//!
+//! Rabin fingerprinting is the default delta engine. Building without `diff_rabin` falls back to
+//! Myers (`diff_mydrs`) or Patience when neither diff feature is enabled.
+//!
+//! Pack bytes are sent through an async channel instead of being written directly. This keeps
+//! object preparation, CPU-heavy compression, and file or network output decoupled. When an index
+//! sender is configured, the encoder also records each object's hash, CRC32, and pack offset so an
+//! `.idx` v2 file can be generated after the pack is complete.
+
+mod delta_search;
+mod header;
+mod parallel;
+mod sort;
+
+pub mod output;
+pub use output::encode_and_output_to_files;
+
+#[cfg(test)]
+mod tests;
+
+use std::{
+    sync::{Arc, Mutex},
+    thread,
+};
+
+use tokio::{sync::mpsc, task::JoinHandle};
+
+use crate::{
+    errors::GitError,
+    hash::ObjectHash,
+    internal::{
+        metadata::{EntryMeta, MetaAttached},
+        object::types::ObjectType,
+        pack::{entry::Entry, index_entry::IndexEntry, pack_index::IdxBuilder},
+    },
+    utils::HashAlgorithm,
+};
+
+use header::encode_header;
+use sort::magic_sort;
+
+/// Stateful encoder for one Git pack stream.
+///
+/// A `PackEncoder` is single-use. It tracks the current pack offset and checksum while encoded
+/// chunks are sent to `pack_sender`. If `idx_sender` is present, index records are accumulated and
+/// can be emitted with [`PackEncoder::encode_idx_file`] after pack encoding finishes.
+pub struct PackEncoder {
+    /// Object count written into the 12-byte pack header.
+    object_number: usize,
+    /// Number of objects consumed by the no-delta path.
+    process_index: usize,
+    /// Zero selects independent encoding; a non-zero value enables delta search.
+    window_size: usize,
+    /// Destination for pack byte chunks. Dropping it signals end-of-stream to the consumer.
+    pack_sender: Option<mpsc::Sender<Vec<u8>>>,
+    /// Optional destination for `.idx` byte chunks.
+    idx_sender: Option<mpsc::Sender<Vec<u8>>>,
+    /// Index metadata collected while pack entries are written.
+    idx_entries: Option<Vec<IndexEntry>>,
+    /// Absolute byte offset where the next pack entry will begin.
+    inner_offset: usize,
+    /// Running checksum of the pack, excluding the checksum trailer itself.
+    inner_hash: HashAlgorithm,
+    /// Final pack checksum, available only after all entries have been encoded.
+    final_hash: Option<ObjectHash>,
+    /// Guards against starting a second encoding operation on the same instance.
+    start_encoding: bool,
+    /// Skip the cheap similarity pre-filter and score every otherwise valid window candidate.
+    ///
+    /// This can improve compression at the cost of more delta computation. The default Rabin path
+    /// disables the pre-filter to follow Git's candidate-search behavior.
+    pub disable_prefilter: bool,
+}
+
+impl PackEncoder {
+    /// Create an encoder that streams pack bytes but does not build an index.
+    pub fn new(object_number: usize, window_size: usize, sender: mpsc::Sender<Vec<u8>>) -> Self {
+        PackEncoder {
+            object_number,
+            window_size,
+            process_index: 0,
+            pack_sender: Some(sender),
+            idx_sender: None,
+            idx_entries: None,
+            inner_offset: 12, // The first entry begins immediately after the pack header.
+            inner_hash: HashAlgorithm::new(),
+            final_hash: None,
+            start_encoding: false,
+            disable_prefilter: false,
+        }
+    }
+
+    /// Create an encoder that streams pack bytes and retains metadata for a later `.idx` stream.
+    pub fn new_with_idx(
+        object_number: usize,
+        window_size: usize,
+        pack_sender: mpsc::Sender<Vec<u8>>,
+        idx_sender: mpsc::Sender<Vec<u8>>,
+    ) -> Self {
+        PackEncoder {
+            object_number,
+            window_size,
+            process_index: 0,
+            pack_sender: Some(pack_sender),
+            idx_sender: Some(idx_sender),
+            idx_entries: None,
+            inner_offset: 12, // The first entry begins immediately after the pack header.
+            inner_hash: HashAlgorithm::new(),
+            final_hash: None,
+            start_encoding: false,
+            disable_prefilter: false,
+        }
+    }
+
+    /// Close the pack output stream by dropping the encoder's final sender.
+    pub fn drop_sender(&mut self) {
+        self.pack_sender.take();
+    }
+
+    /// Send one already-encoded byte chunk to the pack consumer.
+    pub async fn send_data(&mut self, data: Vec<u8>) {
+        if let Some(sender) = &self.pack_sender {
+            sender.send(data).await.unwrap();
+        }
+    }
+
+    /// Return the pack checksum after encoding has completed.
+    pub fn get_hash(&self) -> Option<ObjectHash> {
+        self.final_hash
+    }
+
+    /// Consume an entry stream and emit one complete pack stream.
+    ///
+    /// With `window_size == 0`, entries are independently zlib-compressed in parallel. With a
+    /// non-zero window, entries are collected, sorted for delta locality, and passed through
+    /// sliding-window base selection.
+    ///
+    /// When `diff_rabin` is enabled, the default delta engine is Rabin and all otherwise valid
+    /// candidates are scored. Without that feature, the Myers/Patience engine is used after the
+    /// configured similarity pre-filter.
+    pub async fn encode(
+        &mut self,
+        entry_rx: mpsc::Receiver<MetaAttached<Entry, EntryMeta>>,
+    ) -> Result<(), GitError> {
+        if self.window_size == 0 {
+            self.parallel_encode(entry_rx).await
+        } else {
+            #[cfg(feature = "diff_rabin")]
+            {
+                self.inner_encode(entry_rx, false, true, true).await
+            }
+            #[cfg(not(feature = "diff_rabin"))]
+            {
+                self.inner_encode(entry_rx, false, false, self.disable_prefilter).await
+            }
+        }
+    }
+
+    /// Encode with zstdelta payloads while retaining the same sorting and base-selection pipeline.
+    pub async fn encode_with_zstdelta(
+        &mut self,
+        entry_rx: mpsc::Receiver<MetaAttached<Entry, EntryMeta>>,
+    ) -> Result<(), GitError> {
+        self.inner_encode(entry_rx, true, false, self.disable_prefilter)
+            .await
+    }
+
+    /// Collect, order, delta-search, and write entries for the windowed encoding path.
+    ///
+    /// This method deliberately separates CPU-heavy delta discovery from ordered pack output:
+    ///
+    /// 1. Drain the input channel and partition entries by Git object type.
+    /// 2. Sort each type so likely-related objects are close enough to share a delta window.
+    /// 3. Search for delta bases on blocking worker threads.
+    /// 4. Restore deterministic chunk order and write encoded entries serially, assigning their
+    ///    absolute pack offsets.
+    /// 5. Append the pack checksum and close the output channel.
+    ///
+    /// Candidate-selection rules are based on Git's pack heuristics:
+    /// <https://github.com/git/git/blob/master/Documentation/technical/pack-heuristics.adoc>.
+    async fn inner_encode(
+        &mut self,
+        mut entry_rx: mpsc::Receiver<MetaAttached<Entry, EntryMeta>>,
+        enable_zstdelta: bool,
+        enable_rabin: bool,
+        disable_prefilter: bool,
+    ) -> Result<(), GitError> {
+        // The trailer checksum covers the header and every encoded entry, but not the trailer.
+        let head = encode_header(self.object_number);
+        self.send_data(head.clone()).await;
+        self.inner_hash.update(&head);
+
+        // Reusing the same encoder would corrupt its running offset and checksum state.
+        if self.start_encoding {
+            return Err(GitError::PackEncodeError(
+                "encoding operation is already in progress".to_string(),
+            ));
+        }
+
+        // Delta bases must have the same object type. Keeping types separate makes that invariant
+        // explicit and prevents unrelated types from occupying one another's search windows.
+        let mut commits: Vec<MetaAttached<Entry, EntryMeta>> = Vec::new();
+        let mut trees: Vec<MetaAttached<Entry, EntryMeta>> = Vec::new();
+        let mut blobs: Vec<MetaAttached<Entry, EntryMeta>> = Vec::new();
+        let mut tags: Vec<MetaAttached<Entry, EntryMeta>> = Vec::new();
+        while let Some(entry) = entry_rx.recv().await {
+            match entry.inner.obj_type {
+                ObjectType::Commit => {
+                    commits.push(entry);
+                }
+                ObjectType::Tree => {
+                    trees.push(entry);
+                }
+                ObjectType::Blob => {
+                    blobs.push(entry);
+                }
+                ObjectType::Tag => {
+                    tags.push(entry);
+                }
+                _ => {
+                    return Err(GitError::PackEncodeError(format!(
+                        "object type `{}` is not supported by delta-window pack encoding",
+                        entry.inner.obj_type
+                    )));
+                }
+            }
+        }
+
+        // Sorting is the compression heuristic: nearby entries become eligible delta bases.
+        commits.sort_by(magic_sort);
+        trees.sort_by(magic_sort);
+        blobs.sort_by(magic_sort);
+        tags.sort_by(magic_sort);
+        tracing::info!(
+            "numbers :  commits: {:?} trees: {:?} blobs:{:?} tag :{:?}",
+            commits.len(),
+            trees.len(),
+            blobs.len(),
+            tags.len()
+        );
+
+        // Delta search is CPU-bound, so it must not run on Tokio's async executor threads.
+        // Commits, trees, and tags each get one blocking task. Blobs usually dominate the workload,
+        // so their sorted sequence is divided into contiguous chunks served by a worker queue.
+
+        let num_threads = std::env::var("PACK_THREADS")
+            .ok()
+            .and_then(|s| s.parse::<usize>().ok())
+            .unwrap_or_else(|| {
+                std::thread::available_parallelism()
+                    .map(|n| n.get())
+                    .unwrap_or(4)
+            });
+
+        // Paths are needed only for sorting. Delta generation works on the underlying entries.
+        let commit_entries: Vec<Entry> = commits.into_iter().map(|e| e.inner).collect();
+        let tree_entries: Vec<Entry> = trees.into_iter().map(|e| e.inner).collect();
+        let tag_entries: Vec<Entry> = tags.into_iter().map(|e| e.inner).collect();
+        let mut blob_entries: Vec<Entry> = blobs.into_iter().map(|e| e.inner).collect();
+
+        tracing::info!(
+            blobs = blob_entries.len(),
+            threads = num_threads,
+            "splitting blob delta search"
+        );
+
+        // These object classes are generally small enough that one blocking task per class is
+        // sufficient. Each task owns an independent delta window.
+        let mk_blk = |entries: Vec<Entry>,
+                       ez: bool,
+                       er: bool,
+                       dp: bool|
+         -> tokio::task::JoinHandle<Result<Vec<(Vec<u8>, IndexEntry)>, GitError>> {
+            tokio::task::spawn_blocking(move || {
+                // The search window is currently fixed at 10 inside the parallel delta path.
+                Self::try_as_offset_delta(entries, 10, ez, er, dp)
+            })
+        };
+
+        let commit_handle =
+            mk_blk(commit_entries, enable_zstdelta, enable_rabin, disable_prefilter);
+        let tree_handle = mk_blk(tree_entries, enable_zstdelta, enable_rabin, disable_prefilter);
+        let tag_handle = mk_blk(tag_entries, enable_zstdelta, enable_rabin, disable_prefilter);
+
+        // Preserve contiguous portions of the sorted blob list: splitting objects arbitrarily
+        // would destroy the locality created by magic_sort and reduce delta quality. Creating more
+        // chunks than workers improves load balancing because large blobs cost more to compare.
+        let total_blob_entries = blob_entries.len();
+        let chunks_per_thread: usize = 20;
+        let blob_chunk_count = if num_threads > 1 && total_blob_entries > (num_threads * 20) {
+            num_threads * chunks_per_thread
+        } else {
+            1
+        };
+        let entries_per_chunk = total_blob_entries.div_ceil(blob_chunk_count);
+
+        struct BlobChunk {
+            index: usize,
+            entries: Vec<Entry>,
+        }
+
+        // split_off removes from the end in O(1). Reverse afterward to recover global sort order.
+        let mut chunks: Vec<BlobChunk> = Vec::with_capacity(blob_chunk_count);
+        for i in 0..blob_chunk_count {
+            let take = entries_per_chunk.min(blob_entries.len());
+            if take == 0 {
+                break;
+            }
+            let entries = blob_entries.split_off(blob_entries.len() - take);
+            chunks.push(BlobChunk { index: i, entries });
+        }
+        chunks.reverse(); // Restore original sort order, with larger entries first.
+        for (i, c) in chunks.iter_mut().enumerate() {
+            c.index = i;
+        }
+        let actual_chunks = chunks.len();
+
+        // Workers pop chunks from a shared queue. Results carry their original chunk index because
+        // completion order must not determine pack layout.
+        type ChunkResult = (usize, Result<Vec<(Vec<u8>, IndexEntry)>, GitError>);
+        let chunk_queue: Arc<Mutex<Vec<BlobChunk>>> = Arc::new(Mutex::new(chunks));
+        let results: Arc<Mutex<Vec<ChunkResult>>> = Arc::new(Mutex::new(Vec::new()));
+        let blob_thread_count = num_threads.min(actual_chunks);
+        tracing::info!(
+            total_entries = total_blob_entries,
+            chunks = actual_chunks,
+            entries_per_chunk = entries_per_chunk,
+            threads = blob_thread_count,
+            "splitting blob delta search",
+        );
+        let mut blob_handles: Vec<std::thread::JoinHandle<()>> =
+            Vec::with_capacity(blob_thread_count);
+        for _ in 0..blob_thread_count {
+            let q = Arc::clone(&chunk_queue);
+            let r = Arc::clone(&results);
+            let ez = enable_zstdelta;
+            let er = enable_rabin;
+            let dp = disable_prefilter;
+            blob_handles.push(thread::spawn(move || {
+                loop {
+                    let chunk = q.lock().unwrap().pop();
+                    match chunk {
+                        None => break,
+                        Some(BlobChunk {
+                            index: chunk_idx,
+                            entries,
+                        }) => {
+                            // As above, the per-chunk delta window is currently fixed at 10.
+                            let res =
+                                Self::try_as_offset_delta(entries, 10, ez, er, dp);
+                            r.lock().unwrap().push((chunk_idx, res));
+                        }
+                    }
+                }
+            }));
+        }
+
+        // Join non-blob work and propagate both task failures and encoding failures.
+        let (commit_results, tree_results, tag_results) =
+            tokio::try_join!(commit_handle, tree_handle, tag_handle)
+                .map_err(|e| GitError::PackEncodeError(format!("Task join error: {e}")))?;
+
+        let commit_res = commit_results?;
+        let tree_res = tree_results?;
+        let tag_res = tag_results?;
+
+        // All blob workers must finish before the result Arc can be uniquely unwrapped.
+        for handle in blob_handles {
+            handle.join().map_err(|_| {
+                GitError::PackEncodeError("blob delta thread panicked".to_string())
+            })?;
+        }
+        let mut chunk_results = Arc::into_inner(results)
+            .ok_or_else(|| {
+                GitError::PackEncodeError("blob chunk results still referenced".to_string())
+            })?
+            .into_inner()
+            .map_err(|e| GitError::PackEncodeError(format!("blob results poisoned: {e}")))?;
+        // Parallel search may finish out of order; restore the chosen pack order.
+        chunk_results.sort_by_key(|(i, _)| *i);
+        let mut blob_res_list: Vec<Vec<(Vec<u8>, IndexEntry)>> =
+            Vec::with_capacity(chunk_results.len());
+        for (_idx, res) in chunk_results {
+            blob_res_list.push(res?);
+        }
+
+        let mut all_res = vec![commit_res, tree_res];
+        all_res.extend(blob_res_list);
+        all_res.push(tag_res);
+
+        // Writing is serialized so offsets, the running pack hash, and index records all describe
+        // exactly the same byte order.
+        let mut idx_entries = Vec::new();
+        for res in &mut all_res {
+            for data in res {
+                data.1.offset = self.inner_offset as u64;
+                self.write_all_and_update(&data.0).await;
+                idx_entries.push(data.1.clone());
+            }
+        }
+
+        self.idx_entries = Some(idx_entries);
+
+        // The checksum is both the pack trailer and the identifier used in pack-<hash>.pack.
+        let hash_result = self.inner_hash.clone().finalize();
+        self.final_hash = Some(ObjectHash::from_bytes(&hash_result).unwrap());
+        self.send_data(hash_result.to_vec()).await;
+
+        self.drop_sender();
+        Ok(())
+    }
+
+    /// Account for an encoded chunk, then forward it to the pack consumer.
+    ///
+    /// The caller must invoke this in final pack order because both `inner_offset` and `inner_hash`
+    /// are order-sensitive.
+    async fn write_all_and_update(&mut self, data: &[u8]) {
+        self.inner_hash.update(data);
+        self.inner_offset += data.len();
+        self.send_data(data.to_vec()).await;
+    }
+
+    /// Build the `.idx` stream from metadata captured during pack encoding.
+    async fn generate_idx_file(&mut self) -> Result<(), GitError> {
+        let final_hash = self.final_hash.ok_or(GitError::PackEncodeError(
+            "final_hash is missing,The pack file must be generated before the index file is produced."
+                .into(),
+        ))?;
+        let idx_entries = self.idx_entries.clone().ok_or(GitError::PackEncodeError(
+            "The pack file must be generated before the index file is produced.".into(),
+        ))?;
+        let mut idx_builder = IdxBuilder::new(
+            self.object_number,
+            self.idx_sender.clone().unwrap(),
+            final_hash,
+        );
+        idx_builder.write_idx(idx_entries).await?;
+        Ok(())
+    }
+
+    /// Spawn pack encoding as a Tokio task and return its join handle.
+    ///
+    /// This consumes the encoder so the spawned task owns all channel senders and state. The task
+    /// chooses the same zero-window versus delta-window path as [`PackEncoder::encode`].
+    ///
+    /// Encoding errors currently panic inside the spawned task and are therefore reported as a
+    /// `JoinError` when the returned handle is awaited.
+    pub async fn encode_async(
+        mut self,
+        rx: mpsc::Receiver<MetaAttached<Entry, EntryMeta>>,
+    ) -> Result<JoinHandle<()>, GitError> {
+        Ok(tokio::spawn(async move {
+            if self.window_size == 0 {
+                self.parallel_encode(rx).await.unwrap()
+            } else {
+                self.encode(rx).await.unwrap()
+            }
+        }))
+    }
+
+    /// Spawn zstdelta pack encoding as a Tokio task.
+    ///
+    /// zstdelta always uses the windowed path because independent encoding has no base from which
+    /// to produce a delta.
+    pub async fn encode_async_with_zstdelta(
+        mut self,
+        rx: mpsc::Receiver<MetaAttached<Entry, EntryMeta>>,
+    ) -> Result<JoinHandle<()>, GitError> {
+        Ok(tokio::spawn(async move {
+            self.encode_with_zstdelta(rx).await.unwrap()
+        }))
+    }
+
+    /// Emit the `.idx` stream after the pack has been finalized.
+    ///
+    /// Pack encoding must run first because the index trailer contains the pack checksum and each
+    /// record needs its final absolute pack offset.
+    pub async fn encode_idx_file(&mut self) -> Result<(), GitError> {
+        if self.idx_sender.is_none() {
+            return Err(GitError::PackEncodeError(String::from(
+                "idx sender is none",
+            )));
+        }
+        self.generate_idx_file().await?;
+        // Closing the final sender tells the index writer that no more chunks are coming.
+        self.idx_sender.take();
+        Ok(())
+    }
+}
